Move streaming CompleteDataSetRegistration to streaming sub-package. Add CompleteDataSetRegistrations
