+++ dhis-2/dhis-services/dhis-service-dxf2/src/main/java/org/hisp/dhis/dxf2/dataset/streaming/StreamingXmlCompleteDataSetRegistration.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.dxf2.dataset;
+package org.hisp.dhis.dxf2.dataset.streaming;
 
 /*
  * Copyright (c) 2004-2015, University of Oslo
